Update: hide auto refresh function.
